+++ Acopter32-STM32F4/libraries/AP_HAL_VRBRAIN/AnalogIn.h
@@ -88,7 +88,7 @@ public:
     AP_HAL::AnalogSource* channel(int16_t n);
 
 protected: 
-    VRBRAINAnalogSource* _create_channel(uint8_t num);
+    VRBRAINAnalogSource* _create_channel(int16_t num);
     void _register_channel(VRBRAINAnalogSource*);
     void _timer_event(void);
     VRBRAINAnalogSource* _channels[VRBRAIN_INPUT_MAX_CHANNELS];
